Reland "Android apk-based Gtest: Keep user data dir for PRE_ tests."
This is a reland of commit fa6e7b9cd713f3328ace3b15d010eef0207f295e

Fixes: Removed content/browser/network_service_browsertest.cc change.
Will land this in a separate CL.

Original change's description:
> Android apk-based Gtest: Keep user data dir for PRE_ tests.
>
> https://crrev.com/c/6599562 refactored the code to support the PRE_
> test grouping in gtest. This CL is a followup to keep user data dir
> for PRE_ tests in apk-based gtest.
>
> **Note** that PRE_ tests in android_browsertests still don't work
> because it uses a different way to setup and clean user data dir,
> after the CL https://crrev.com/c/1677591
>
> This CL introduces new intent extra "KeepUserDataDir" in
> 1. NativeTestInstrumentationTestRunner.
> 2. NativeTestIntent.
>
> The python code local_device_gtest_run.py sets the extra #1 to true
> for a PRE_ test group.
>
> Then NativeTestInstrumentationTestRunner.java consumes this info and
> sets the extra #2 for each individual java test shard.
>
> The *BrowserTestActivity.java then read this info, and keep the user
> data dir when needed, before running the tests.
>
> This CL also enables NetworkServiceBrowserCacheResetTest in
> content_browsertests on android, which contains PRE_ tests
>
> Bug: 40200835
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6568919
> Commit-Queue: Haiyang Pan <hypan@google.com>
> Cr-Commit-Position: refs/heads/main@{#1473200}

Bug: 40200835, 424625656
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6643463
Commit-Queue: Haiyang Pan <hypan@google.com>
Cr-Commit-Position: refs/heads/main@{#1474455}

+++ testing/android/native_test/java/src/org/chromium/native_test/NativeBrowserTestActivity.java
@@ -13,7 +13,7 @@ import java.io.File;
 
 /** An {@link android.app.Activity} for running native browser tests. */
 public abstract class NativeBrowserTestActivity extends FragmentActivity {
-    private static final String TAG = "NativeTest";
+    private static final String TAG = "NativeTestActivity";
 
     private final NativeTest mTest = new NativeTest();
     private boolean mStarted;
@@ -45,7 +45,9 @@ public abstract class NativeBrowserTestActivity extends FragmentActivity {
         if (mStarted) return;
 
         mStarted = true;
-        NativeBrowserTest.deletePrivateDataDirectory(getPrivateDataDirectory());
+        if (!mTest.shouldKeepUserDataDir()) {
+            NativeBrowserTest.deletePrivateDataDirectory(getPrivateDataDirectory());
+        }
         initializeBrowserProcess();
     }
